requirements.txt
line-bot-sdk
flask
pyzbar
Pillow